+++ assets/js/modules/search-console/components/dashboard/SearchFunnelWidget/Overview/index.js
@@ -1,7 +1,7 @@
 /**
  * Overview component for SearchFunnelWidget.
  *
- * Site Kit by Google, Copyright 2023 Google LLC
+ * Site Kit by Google, Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,17 +22,50 @@
 import PropTypes from 'prop-types';
 
 /**
+ * WordPress dependencies
+ */
+import { __ } from '@wordpress/i18n';
+
+/**
  * Internal dependencies
  */
+import Data from 'googlesitekit-data';
 import { ProgressBar } from 'googlesitekit-components';
-import { Grid, Row, Cell } from '../../../../../../material-components';
-import ActivateAnalyticsCTA from '../ActivateAnalyticsCTA';
-import CreateGoalCTA from '../CreateGoalCTA';
-import DataBlock from '../../../../../../components/DataBlock';
-import RecoverableModules from '../../../../../../components/RecoverableModules';
-import { BREAKPOINT_SMALL } from '../../../../../../hooks/useBreakpoint';
-import { useOverviewData } from './useOverviewData';
+import { Grid, Row, Cell } from '../../../../../material-components';
+import { extractSearchConsoleDashboardData } from '../../../util';
+import { calculateChange } from '../../../../../util';
+import { CORE_MODULES } from '../../../../../googlesitekit/modules/datastore/constants';
+import { MODULES_ANALYTICS } from '../../../../analytics/datastore/constants';
+import { CORE_LOCATION } from '../../../../../googlesitekit/datastore/location/constants';
+import { MODULES_SEARCH_CONSOLE } from '../../../datastore/constants';
+import useDashboardType, {
+	DASHBOARD_TYPE_MAIN,
+	DASHBOARD_TYPE_ENTITY,
+} from '../../../../../hooks/useDashboardType';
+import ActivateAnalyticsCTA from './ActivateAnalyticsCTA';
+import CreateGoalCTA from './CreateGoalCTA';
+import DataBlock from '../../../../../components/DataBlock';
+import RecoverableModules from '../../../../../components/RecoverableModules';
+import {
+	BREAKPOINT_SMALL,
+	useBreakpoint,
+} from '../../../../../hooks/useBreakpoint';
+import useViewOnly from '../../../../../hooks/useViewOnly';
+import { CORE_USER } from '../../../../../googlesitekit/datastore/user/constants';
+const { useSelect, useInViewSelect } = Data;
 
+function getDatapointAndChange( [ report ], selectedStat, divider = 1 ) {
+	return {
+		datapoint:
+			report?.data?.totals?.[ 0 ]?.values?.[ selectedStat ] / divider,
+		change: calculateChange(
+			report?.data?.totals?.[ 1 ]?.values?.[ selectedStat ],
+			report?.data?.totals?.[ 0 ]?.values?.[ selectedStat ]
+		),
+	};
+}
+
+// eslint-disable-next-line complexity
 const Overview = ( {
 	analyticsData,
 	analyticsGoalsData,
@@ -45,30 +78,204 @@ const Overview = ( {
 	WidgetReportError,
 	showRecoverableAnalytics,
 } ) => {
-	const {
-		analyticsModuleActive,
-		analyticsModuleActiveAndConnected,
-		analyticsModuleConnected,
-		breakpoint,
-		canViewSharedAnalytics,
-		dataBlockCellProps,
-		dataBlocks,
-		dataBlockWrapperCellProps,
-		halfCellProps,
-		isNavigatingToReauthURL,
-		isSearchConsoleGatheringData,
-		quarterCellProps,
-		showAnalytics,
-		showGoalsCTA,
-	} = useOverviewData( {
-		analyticsData,
-		analyticsGoalsData,
-		analyticsVisitorsData,
-		dateRangeLength,
-		error,
-		searchConsoleData,
-		showRecoverableAnalytics,
+	const dashboardType = useDashboardType();
+	const breakpoint = useBreakpoint();
+
+	const viewOnly = useViewOnly();
+
+	const analyticsModuleAvailable = useSelect( ( select ) =>
+		select( CORE_MODULES ).isModuleAvailable( 'analytics' )
+	);
+
+	const canViewSharedAnalytics = useSelect( ( select ) => {
+		if ( ! analyticsModuleAvailable ) {
+			return false;
+		}
+
+		if ( ! viewOnly ) {
+			return true;
+		}
+
+		return select( CORE_USER ).canViewSharedModule( 'analytics' );
 	} );
+
+	const analyticsModuleConnected = useSelect( ( select ) =>
+		select( CORE_MODULES ).isModuleConnected( 'analytics' )
+	);
+	const analyticsModuleActive = useSelect( ( select ) =>
+		select( CORE_MODULES ).isModuleActive( 'analytics' )
+	);
+	const analyticsModuleActiveAndConnected =
+		analyticsModuleActive && analyticsModuleConnected;
+
+	const isNavigatingToReauthURL = useSelect( ( select ) => {
+		if ( ! analyticsModuleAvailable ) {
+			return false;
+		}
+		const adminReauthURL = select( MODULES_ANALYTICS ).getAdminReauthURL();
+		return select( CORE_LOCATION ).isNavigatingTo( adminReauthURL );
+	} );
+	const isSearchConsoleGatheringData = useInViewSelect( ( select ) =>
+		select( MODULES_SEARCH_CONSOLE ).isGatheringData()
+	);
+	const isAuthenticated = useSelect( ( select ) =>
+		select( CORE_USER ).isAuthenticated()
+	);
+
+	const {
+		totalClicks,
+		totalImpressions,
+		totalClicksChange,
+		totalImpressionsChange,
+	} = extractSearchConsoleDashboardData( searchConsoleData, dateRangeLength );
+
+	let analyticsGoalsChange = null;
+	let analyticsGoalsDatapoint = null;
+	let analyticsBounceDatapoint = null;
+	let analyticsBounceChange = null;
+	let analyticsVisitorsDatapoint = null;
+	let analyticsVisitorsChange = null;
+
+	if (
+		analyticsModuleActive &&
+		Array.isArray( analyticsData ) &&
+		Array.isArray( analyticsVisitorsData )
+	) {
+		( { change: analyticsGoalsChange } = getDatapointAndChange(
+			analyticsData,
+			0,
+			100
+		) );
+		analyticsGoalsDatapoint =
+			analyticsData?.[ 0 ]?.data?.totals?.[ 0 ]?.values[ 0 ];
+
+		( {
+			datapoint: analyticsBounceDatapoint,
+			change: analyticsBounceChange,
+		} = getDatapointAndChange( analyticsData, 1, 100 ) );
+
+		analyticsVisitorsDatapoint =
+			analyticsVisitorsData?.[ 0 ]?.data?.totals?.[ 0 ]?.values[ 0 ];
+		( { change: analyticsVisitorsChange } = getDatapointAndChange(
+			analyticsVisitorsData,
+			0,
+			100
+		) );
+	}
+
+	const showAnalytics =
+		canViewSharedAnalytics &&
+		analyticsModuleConnected &&
+		! error &&
+		! showRecoverableAnalytics;
+
+	const showGoalsCTA =
+		isAuthenticated &&
+		showAnalytics &&
+		dashboardType === DASHBOARD_TYPE_MAIN &&
+		! analyticsGoalsData?.items?.length;
+
+	const quarterCellProps = {
+		smSize: 2,
+		mdSize: showGoalsCTA ? 4 : 2,
+		lgSize: 3,
+	};
+
+	const oneThirdCellProps = {
+		smSize: 2,
+		mdSize: 4,
+		lgSize: 4,
+	};
+
+	const halfCellProps = {
+		smSize: 4,
+		mdSize: 4,
+		lgSize: 6,
+	};
+
+	const threeQuartersCellProps = {
+		smSize: 4,
+		mdSize: 4,
+		lgSize: 9,
+	};
+
+	const fullCellProps = {
+		smSize: 4,
+		mdSize: 8,
+		lgSize: 12,
+	};
+
+	// Collection of all the data blocks to be displayed
+	const dataBlocks = [
+		{
+			id: 'impressions',
+			stat: 0,
+			title: __( 'Total Impressions', 'google-site-kit' ),
+			datapoint: totalImpressions,
+			change: totalImpressionsChange,
+		},
+		{
+			id: 'clicks',
+			stat: 1,
+			title: __( 'Total Clicks', 'google-site-kit' ),
+			datapoint: totalClicks,
+			change: totalClicksChange,
+		},
+		...( showAnalytics
+			? [
+					{
+						id: 'visitors',
+						stat: 2,
+						title: __(
+							'Unique Visitors from Search',
+							'google-site-kit'
+						),
+						datapoint: analyticsVisitorsDatapoint,
+						change: analyticsVisitorsChange,
+					},
+			  ]
+			: [] ),
+		...( showAnalytics &&
+		dashboardType === DASHBOARD_TYPE_MAIN &&
+		analyticsGoalsData?.items?.length > 0
+			? [
+					{
+						id: 'goals',
+						stat: 3,
+						title: __( 'Goals', 'google-site-kit' ),
+						datapoint: analyticsGoalsDatapoint,
+						change: analyticsGoalsChange,
+					},
+			  ]
+			: [] ),
+		...( showAnalytics && dashboardType === DASHBOARD_TYPE_ENTITY
+			? [
+					{
+						id: 'bounce',
+						stat: 4,
+						title: __( 'Bounce Rate', 'google-site-kit' ),
+						datapoint: analyticsBounceDatapoint,
+						datapointUnit: '%',
+						change: analyticsBounceChange,
+					},
+			  ]
+			: [] ),
+	];
+
+	const dataBlockWrapperCellProps = {
+		2: halfCellProps,
+		3: threeQuartersCellProps,
+		4: fullCellProps,
+	};
+
+	const dataBlockCellProps = {
+		2: {
+			...halfCellProps,
+			smSize: 2,
+		},
+		3: oneThirdCellProps,
+		4: quarterCellProps,
+	};
 
 	return (
 		<Grid>
